+++ src/plone/restapi/services/content/get.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from AccessControl.interfaces import IRoleManager
+
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.services import Service
 from zope.component import queryMultiAdapter
@@ -16,4 +18,10 @@ class ContentGet(Service):
             self.request.response.setStatus(501)
             return dict(error=dict(message='No serializer available.'))
 
-        return serializer()
+        data = serializer()
+        if IRoleManager.providedBy(self.context):
+            data['sharing'] = {
+                '@id': '{}/@sharing'.format(self.context.absolute_url()),
+                'title': 'Sharing',
+                }
+        return data