+++ src/plone/restapi/tests/test_documentation.py
@@ -446,7 +446,6 @@ class TestTraversal(unittest.TestCase):
         response = self.api_session.post(
             '/@users',
             json={
-                'username': 'noam',
                 'email': 'noam.chomsky@example.com',
                 'password': 'colorlessgreenideas',
                 'username': 'noamchomsky',
@@ -651,10 +650,10 @@ class TestTraversal(unittest.TestCase):
 
     def test_documentation_vocabularies_all(self):
         response = self.api_session.get('/@vocabularies')
-        save_request_and_response_for_docs('vocabularies.json', response)
+        save_request_and_response_for_docs('vocabularies', response)
 
     def test_documentation_vocabularies_get(self):
         response = self.api_session.get(
             '/@vocabularies/plone.app.vocabularies.ReallyUserFriendlyTypes'
         )
-        save_request_and_response_for_docs('vocabularies_get.json', response)
+        save_request_and_response_for_docs('vocabularies_get', response)