+++ src/plone/restapi/tests/test_services_users.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+from Products.CMFCore.permissions import SetOwnPassword
+from Products.CMFCore.utils import getToolByName
+from Products.MailHost.interfaces import IMailHost
+
 from plone import api
 from plone.app.testing import setRoles
 from plone.app.testing import SITE_OWNER_NAME
@@ -7,6 +11,7 @@ from plone.app.testing import TEST_USER_ID
 from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
 from plone.restapi.testing import RelativeSession
 from zope.component import getAdapter
+from zope.component import getUtility
 
 try:
     from Products.CMFPlone.interfaces import ISecuritySchema
@@ -27,9 +32,13 @@ class TestUsersEndpoint(unittest.TestCase):
         self.portal_url = self.portal.absolute_url()
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
 
+        self.mailhost = getUtility(IMailHost)
+
         self.api_session = RelativeSession(self.portal_url)
         self.api_session.headers.update({'Accept': 'application/json'})
         self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+        self.anon_api_session = RelativeSession(self.portal_url)
+        self.anon_api_session.headers.update({'Accept': 'application/json'})
 
         properties = {
             'email': 'noam.chomsky@example.com',
@@ -42,7 +51,8 @@ class TestUsersEndpoint(unittest.TestCase):
         api.user.create(
             email='noam.chomsky@example.com',
             username='noam',
-            properties=properties
+            properties=properties,
+            password=u'password'
         )
         transaction.commit()
 
@@ -94,7 +104,7 @@ class TestUsersEndpoint(unittest.TestCase):
         transaction.commit()
 
         self.assertEqual(400, response.status_code)
-        self.assertTrue('"Property \'username\' is required' in response.text)
+        self.assertTrue('Property \'username\' is required' in response.text)
 
     def test_add_user_password_is_required(self):
         response = self.api_session.post(
@@ -106,7 +116,9 @@ class TestUsersEndpoint(unittest.TestCase):
         transaction.commit()
 
         self.assertEqual(400, response.status_code)
-        self.assertTrue('"Property \'password\' is required' in response.text)
+        self.assertTrue(
+            ('You have to either send a '
+             'password or sendPasswordReset') in response.text)
 
     def test_add_user_email_is_required_if_email_login_is_enabled(self):
         # enable use_email_as_login
@@ -122,7 +134,8 @@ class TestUsersEndpoint(unittest.TestCase):
         )
 
         self.assertEqual(400, response.status_code)
-        self.assertTrue('"Property \'email\' is required' in response.text)
+        self.assertTrue('Property \'username\' is not allowed' in
+                        response.text)
 
     def test_add_user_email_with_email_login_enabled(self):
         # enable use_email_as_login
@@ -141,7 +154,7 @@ class TestUsersEndpoint(unittest.TestCase):
         self.assertEqual(201, response.status_code)
         self.assertTrue(api.user.get(userid='howard.zinn@example.com'))
 
-    def test_add_user_with_email_login_enabled(self):
+    def test_username_is_not_allowed_with_email_login_enabled(self):
         # enable use_email_as_login
         security_settings = getAdapter(self.portal, ISecuritySchema)
         security_settings.use_email_as_login = True
@@ -156,11 +169,79 @@ class TestUsersEndpoint(unittest.TestCase):
         )
         transaction.commit()
 
+        self.assertEqual(400, response.status_code)
+        self.assertTrue('Property \'username\' is not allowed'
+                        in response.text)
+
+    def test_add_user_with_email_login_enabled(self):
+        # enable use_email_as_login
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.use_email_as_login = True
+        transaction.commit()
+        response = self.api_session.post(
+            '/@users',
+            json={
+                "email": "howard.zinn@example.com",
+                "password": "secret"
+            },
+        )
+        transaction.commit()
+
         self.assertEqual(201, response.status_code)
         user = api.user.get(userid='howard.zinn@example.com')
         self.assertTrue(user)
         self.assertEqual('howard.zinn@example.com', user.getUserName())
         self.assertEqual('howard.zinn@example.com', user.getProperty('email'))
+
+    def test_add_user_with_sendPasswordRest_sends_email(self):
+        response = self.api_session.post(
+            '/@users',
+            json={
+                "username": "howard",
+                "email": "howard.zinn@example.com",
+                "sendPasswordReset": True
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+        self.assertTrue('To: howard.zinn@example.com' in
+                        self.mailhost.messages[0])
+
+    def test_add_user_send_properties(self):
+        response = self.api_session.post(
+            '/@users',
+            json={
+                "username": "howard",
+                "password": "secret",
+                "email": "howard.zinn@example.com",
+                "fullname": "Howard Zinn",
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+        member = api.user.get(username='howard')
+        self.assertEqual(member.getProperty('fullname'), 'Howard Zinn')
+
+    def test_add_anon_user_sends_properties_are_not_saved(self):
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.enable_self_reg = True
+        transaction.commit()
+
+        response = self.anon_api_session.post(
+            '/@users',
+            json={
+                "username": "howard",
+                "email": "howard.zinn@example.com",
+                "fullname": "Howard Zinn",
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+        member = api.user.get(username='howard')
+        self.assertEqual(member.getProperty('fullname'), '')
 
     def test_get_user(self):
         response = self.api_session.get('/@users/noam')
@@ -249,6 +330,91 @@ class TestUsersEndpoint(unittest.TestCase):
             old_password_hashes['noam'], new_password_hashes['noam']
         )
 
+    def test_user_requests_password_reset_mail(self):
+        self.api_session.auth = ('noam', 'password')
+        payload = {}
+        response = self.api_session.post('/@users/noam/reset-password',
+                                         json=payload)
+        transaction.commit()
+
+        self.assertEqual(response.status_code, 200)
+        # FIXME: Test that mail is sent
+
+    def test_user_set_own_password(self):
+        self.api_session.auth = ('noam', 'password')
+        self.portal.manage_permission(
+            SetOwnPassword, roles=['Authenticated', 'Manager'], acquire=False)
+        transaction.commit()
+
+        payload = {'old_password': 'password',
+                   'new_password': 'new_password'}
+        response = self.api_session.post('/@users/noam/reset-password',
+                                         json=payload)
+        transaction.commit()
+
+        self.assertEqual(response.status_code, 200)
+        authed = self.portal.acl_users.authenticate('noam', 'new_password',
+                                                    {})
+        self.assertTrue(authed)
+
+    def test_user_set_own_password_requires_set_own_password_permission(self):
+        self.api_session.auth = ('noam', 'password')
+        self.portal.manage_permission(SetOwnPassword, roles=['Manager'],
+                                      acquire=False)
+        transaction.commit()
+
+        payload = {'old_password': 'password',
+                   'new_password': 'new_password'}
+        response = self.api_session.post('/@users/noam/reset-password',
+                                         json=payload)
+        transaction.commit()
+
+        self.assertEqual(response.status_code, 403)
+
+    def test_user_set_own_password_requires_old_and_new_password(self):
+        self.api_session.auth = ('noam', 'password')
+        payload = {'old_password': 'password'}
+        response = self.api_session.post('/@users/noam/reset-password',
+                                         json=payload)
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()['error']['type'],
+                         'Invalid parameters')
+        payload = {'new_password': 'new_password'}
+        response = self.api_session.post('/@users/noam/reset-password',
+                                         json=payload)
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()['error']['type'],
+                         'Invalid parameters')
+
+    def test_user_set_reset_token_requires_new_password(self):
+        self.api_session.auth = ('noam', 'password')
+        payload = {'reset_token': 'abc'}
+        response = self.api_session.post('/@users/noam/reset-password',
+                                         json=payload)
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()['error']['type'],
+                         'Invalid parameters')
+
+    def test_reset_with_token(self):
+        reset_tool = getToolByName(self.portal, 'portal_password_reset')
+        reset_info = reset_tool.requestReset('noam')
+        token = reset_info['randomstring']
+        transaction.commit()
+
+        payload = {'reset_token': token,
+                   'new_password': 'new_password'}
+        response = self.api_session.post('/@users/noam/reset-password',
+                                         json=payload)
+        transaction.commit()
+
+        self.assertEqual(response.status_code, 200)
+        authed = self.portal.acl_users.authenticate('noam', 'new_password',
+                                                    {})
+        self.assertTrue(authed)
+
     def test_delete_user(self):
         response = self.api_session.delete('/@users/noam')
         transaction.commit()
@@ -261,3 +427,103 @@ class TestUsersEndpoint(unittest.TestCase):
         transaction.commit()
 
         self.assertEqual(response.status_code, 404)
+
+    def test_anonymous_requires_enable_self_reg(self):
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.enable_self_reg = False
+        transaction.commit()
+
+        response = self.anon_api_session.post(
+            '/@users',
+            json={
+                "password": "noamchomsky"
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(403, response.status_code)
+
+        security_settings.enable_self_reg = True
+        transaction.commit()
+
+        response = self.anon_api_session.post(
+            '/@users',
+            json={
+                "username": "new_user",
+                'email': 'avram.chomsky@example.com'
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+
+    def test_anonymous_without_enable_user_pwd_choice_sends_mail(self):
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.enable_self_reg = True
+        transaction.commit()
+
+        response = self.anon_api_session.post(
+            '/@users',
+            json={
+                "username": "new_user",
+                'email': 'avram.chomsky@example.com'
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+        self.assertTrue('To: avram.chomsky@example.com' in
+                        self.mailhost.messages[0])
+
+    def test_anonymous_can_set_password_with_enable_user_pwd_choice(self):
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.enable_self_reg = True
+        transaction.commit()
+
+        response = self.anon_api_session.post(
+            '/@users',
+            json={
+                "username": "new_user",
+                'email': 'avram.chomsky@example.com',
+                'password': 'secret'
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(400, response.status_code)
+        self.assertTrue('Property \'password\' is not allowed'
+                        in response.text)
+
+        security_settings.enable_user_pwd_choice = True
+        transaction.commit()
+
+        response = self.anon_api_session.post(
+            '/@users',
+            json={
+                "username": "new_user",
+                'email': 'avram.chomsky@example.com',
+                'password': 'secret'
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+
+    def test_anonymous_with_enable_user_pwd_choice_doent_send_email(self):
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.enable_self_reg = True
+        security_settings.enable_user_pwd_choice = True
+        transaction.commit()
+
+        response = self.anon_api_session.post(
+            '/@users',
+            json={
+                "username": "new_user",
+                'email': 'avram.chomsky@example.com',
+                'password': 'secret'
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(self.mailhost.messages, [])
+        self.assertEqual(201, response.status_code)